Continue early work; restructurings
I moved the code to use C++ namespaces so that we may use more general
names in the code. This has the side effect of being horribly verbose at
times. C++ is a bad language, simple as.

I prepared some code for the taskbar proper, but it's terribly unfinished
for now. I'm checking in what I've got so far so that the other developers
can take a look at it.

Signed-off-by: Isabella <kawapure@gmail.com>