[CTaskbar] Hide multiple taskbars; improve hiding
There is now code to hide multiple taskbars, and I experimented a bit with
adding a timer to fight with Explorer's taskbar a little bit. It doesn't
seem to work that well, unfortunately. It seems to slow down my computer
(not just the KappaBar test window) quite a bit, and it still seems to let
the official taskbar come in sometimes.

I speculate that this may be the case because we haven't installed a window
with the same class name as the official taskbar yet. Once that's done, it
probably stops fighting. That's probably why such performance issues aren't
really observed from, for example, RetroBar (which uses ManagedShell).

Overall, I copied the hiding procedure from ManagedShell because it seems
to be required in order to make the taskbar stay hidden, at least before a
custom taskbar window can take over. It looks like ManagedShell leaves it
running in the background forever, which could warrant optimisation if we
must do the same.

Signed-off-by: Isabella <kawapure@gmail.com>