[Util] Add a versioning system There are two version numbers; there's the version number which is set by us manually, and there's the build number which is calculated at compile time.
The build number calculation is identical to the calculation for the
Source Engine (pre-2013). We take the date at compile time (__DATE__),
convert it into a number of days, and then subtract a certain epoch date
from it. For the Source Engine, this was Nov 07 1998, 12 days before the
release of the original Half-Life, described by Source as the "HL1 Gold
Date". For KappaBar, I changed this to Jan 08 2024, which is the date of
the initial commit to the repository.

Luckily, I managed to cram all of this build number calculation into the
compile time, unlike the Source Engine, which just has a date string to
compile the build number at runtime from.